requirements.txt
# Core ML Dependencies
torch==2.0.1
torchvision==0.15.2
torchxrayvision==1.2.0
captum==0.6.0

# Image Processing
pillow==10.0.0
opencv-python==4.8.1
scikit-image==0.21.0
imageio==2.31.1

# Scientific Computing
numpy==1.24.3
scipy==1.11.2
pandas==2.0.3
scikit-learn==1.3.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# GUI
tk==8.6.12

# PDF Generation
reportlab==4.0.4

# Medical Imaging
pydicom==2.4.3
nibabel==5.1.0
SimpleITK==2.3.0

# Progress Bars
tqdm==4.66.1

# Configuration
pyyaml==6.0.1
json5==0.9.14

# Logging
loguru==0.7.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Code Quality
black==23.7.0
flake8==6.1.0
mypy==1.5.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
